src/third_party/WebKit/LayoutTests/external/wpt/css/css-box/parsing/padding-invalid-expected.txt

This is a testharness.js-based test.
PASS e.style['padding'] = "auto" should not set the property value
PASS e.style['padding'] = "available" should not set the property value
PASS e.style['padding'] = "10px border-box" should not set the property value
PASS e.style['padding'] = "calc(2em + 3ex) auto" should not set the property value
PASS e.style['padding'] = "10px 20px 30px 40px 50px" should not set the property value
PASS e.style['padding-top'] = "auto" should not set the property value
PASS e.style['padding-bottom'] = "10px 20px calc(2em + 3ex) auto" should not set the property value
PASS e.style['padding-bottom-left'] = "10px 20px" should not set the property value
PASS e.style['padding-top'] = "-10px" should not set the property value
FAIL e.style['padding'] = "20%" should not set the property value assert_equals: expected "" but got "20%"
FAIL e.style['padding'] = "10px 20% 30% 40px" should not set the property value assert_equals: expected "" but got "10px 20% 30% 40px"
FAIL e.style['padding-right'] = "20%" should not set the property value assert_equals: expected "" but got "20%"
FAIL e.style['padding-right'] = "calc(2em + 3%)" should not set the property value assert_equals: expected "" but got "calc(2em + 3%)"
Harness: the test ran to completion.
